requirements.txt
selenium==3.13.0
Django==2.0.7
django-extensions==2.1.0
# mysqlclient==1.3.13
PyMySQL==0.9.2
djangorestframework==3.8.2
xlrd==1.1.0
xlwt==1.3.0
requests==2.19.1
zeep==3.1.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
xlsxwriter==1.0.5
eleme.openapi.python.sdk==1.18.5
curlify==2.0.1
drf-nested-routers==0.90.2
openpyxl==2.5.4
pyexcel-xls==0.5.7
pyexcel-xlsx==0.5.6
django-filter==2.0.0
django-crontab==0.7.1
django-debug-toolbar==1.9.1
inflection==0.3.1
djangorestframework-csv==2.1.0
jsonfield==2.0.2
retrying==1.3.3
aliyun-python-sdk-core-v3==2.9.1
aliyun-python-sdk-dysmsapi==1.0.0
redis==2.10.6
django-redis==4.9.0
dingtalk-sdk==1.1.5
django-mysql==2.4.1
dulwich==0.19.6
gevent==1.3.6
chinesecalendar
https://github.com/celery/celery/tarball/5cdaeb
# git+https://github.com/gikoluo/aliyun-python-sdk-core.git
aliyun-python-sdk-iot
djangorestframework-xml==1.3.0
docxtpl==0.5.8
python-dateutil==2.7.5
wechatpy==1.7.5
django-guardian==1.4.9
rest-framework-generic-relations==1.2.1
pyenchant==2.0.0
raven==6.10.0
elastic-apm==4.0.3
dicttoxml==1.7.4
Jinja2==2.10
python-redis-lock[django]==3.3.1
pypinyin==0.35.1
python-intervals==1.8.0
django-cacheops==4.1
django-sentry-400-middleware==0.3.0
croniter==0.3.30
